LayoutTests/platform/mac/accessibility/aria-grouping-roles-expected.txt

application role
article role
banner role
complementary role
content info role
document role
log role
main role
marquee role
note role
navigation role
region role
search role
status role
tooltip role
definition role
This tests that each of these ARIA grouping type roles map to the right role and subrole.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkApplication"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXDocumentArticle"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkBanner"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkComplementary"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkContentInfo"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXDocument"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXApplicationLog"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkMain"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXApplicationMarquee"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXDocumentNote"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkNavigation"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXDocumentRegion"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXLandmarkSearch"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXApplicationStatus"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).subrole is "AXSubrole: AXUserInterfaceTooltip"
PASS obj.childAtIndex(k).role is 'AXRole: AXGroup'
PASS obj.childAtIndex(k).description is 'AXDescription: definition'
PASS successfullyParsed is true

TEST COMPLETE
